DEV: Ajustar la visibilidad de los productos según el estado de visibildiad de la herencia - No permitir eliminar elementos que estan asignados sobre otros
Add AutoMapper configurations and new view models

This commit introduces new AutoMapper mappings for `vwGetAllProducts`, `vwGetAllCategories`, and `vwGetAllSubcategories`, enhancing the DTOs with additional properties. The `CategoryDto` and `SubcategoryDto` classes are updated to include `HasSubcategory` and `HasProduct` boolean properties. The `ProductDto` class's `Price` property type is changed from `float` to `double` for better precision.

New view classes are created in the `DigiMenuAPI.Infrastructure.Entities.Views` namespace, and the `ApplicationDbContext` is updated to include DbSet properties for these views. Migration files are added to create the new views in the database, ensuring the schema is in sync with the updated model. The project file is also updated to include necessary compile and content items related to the new views.